+++ src/store/index.js
@@ -1,0 +1,4 @@
+import { createStore } from 'redux'
+import reducer from './reducers' 
+const store = createStore( reducer )
+export default store
