+++ src/store/type.js
@@ -1,0 +1,5 @@
+const type=
+{
+    DELETE:1
+}
+export default type
